Add files via upload
pyvis is a good tool for visualizing graph structures( https://pyvis.readthedocs.io/en/latest/index.html ).  In the application case of the CPP-LGI algorithm, pyvis is used to visualize the graph structure obtained by discrete sampling. Please use a browser to open 'CPP_LGI_black.html', such as Chrome, Firefox, Edge, etc. The nodes in the graph correspond to the indicators in 'https://github.com/With-the-sun/PN-CRAR-mode-s-casedata.git '